chore: replace old packages
chore: remove lodash & lodash.sortby

Remove packages lodash & lodash.sortby because they were not used in the code.

chore: replace lodash.isequal with just-compare

Replace lodash.isequal with just-compare. Just-compare is 3-5 times faster,
weighs 600B instead of 10KB and has built-in type support.

chore: replace is-empty-object with just-is-epmty

Replace is-empty-object with just-is-empty. Just-is-empty is 1.2-2.5 times faster
and has built-in type support.

chore: replace flatten with just-flatten-it

Replace flatten with just-flatten-it. Just-compare is 1.5-5 times faster
and has built-in type support.